% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation_range.R
\name{apply_circle_perturbation_range}
\alias{apply_circle_perturbation_range}
\title{Circle perturbation range}
\usage{
apply_circle_perturbation_range(pts, r)
}
\arguments{
\item{pts}{\code{matrix} of original points.}

\item{r}{\code{vector} of radii.}
}
\value{
A \code{list} of \code{matrix} objects with perturbated points.
}
\description{
Applies a linear circle perturbation function for a range of parameters to a
set of points.
}
