% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation_range.R
\name{apply_donut_perturbation_range}
\alias{apply_donut_perturbation_range}
\title{Donut perturbation range}
\usage{
apply_donut_perturbation_range(pts, r1, r2)
}
\arguments{
\item{pts}{\code{matrix} of original points.}

\item{r1}{\code{vector} of inner radii.}

\item{r2}{\code{vector} of outer radii.}
}
\value{
A \code{list} of \code{matrix} objects with perturbated points.
}
\description{
Applies a donut perturbation function for a range of parameters to a set of
points.
}
