% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation_range.R
\name{apply_perturbation_range}
\alias{apply_perturbation_range}
\title{Perturbation Range}
\usage{
apply_perturbation_range(pts, perturbation_method, r1, r2)
}
\arguments{
\item{pts}{\code{matrix} of original points.}

\item{perturbation_method}{Method used for perturbation (One of "circle",
"donut", "circle_uniform", "circle_gauss").}

\item{r1}{\code{vector} of radii (or standard deviations for "circle_gauss").}

\item{r2}{\code{vector} of radii (only needed if using donut pertubation method).}
}
\value{
A \code{list} of perturbated points.
}
\description{
Applies a given perturbation function for a range of parameters to a set of
points
}
