% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{circle}
\alias{circle}
\title{Circle line perturbation}
\usage{
circle(p, r)
}
\arguments{
\item{p}{Original point.}

\item{r}{Perturbation radius.}
}
\value{
Perturbated point.
}
\description{
Adds perturbation constrained on a line with fixed length around a point.
}
