% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{circle_gauss}
\alias{circle_gauss}
\title{Gauss Circle}
\usage{
circle_gauss(p, sd)
}
\arguments{
\item{p}{Original point.}

\item{sd}{Standard deviation of the distribution.}
}
\value{
Perturbated point.
}
\description{
Adds perturbation within a circle with normal (Gauss) distribution.
}
