% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{circle_uniform}
\alias{circle_uniform}
\title{Uniform Circle}
\usage{
circle_uniform(p, r)
}
\arguments{
\item{p}{Original point.}

\item{r}{Circle radius.}
}
\value{
Perturbated point.
}
\description{
Adds perturbation within a circle with uniform distribution.
}
