% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{comparison_plots}
\alias{comparison_plots}
\title{Comparison plots}
\usage{
comparison_plots(dat_orig, dat_dist, labels)
}
\arguments{
\item{dat_orig}{\code{data.frame} of original point data.}

\item{dat_dist}{\code{list} containing all data.}

\item{labels}{\code{vector} of perturbation values as labels for the plots.}
}
\description{
Generate plot overlaying the original and distorted data.
}
