% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Calculate descriptive statistics}
\usage{
descriptive_stats(dat, range)
}
\arguments{
\item{dat}{input data.}

\item{range}{\code{vector} of perturbation values.}
}
\value{
A \code{data.frame} containing all statistics.
}
\description{
Calculate the mean, median, standard deviation of a set of perturbated data.
}
