% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{distance_to_zero}
\alias{distance_to_zero}
\title{Distance to zero}
\usage{
distance_to_zero(dat)
}
\arguments{
\item{dat}{\code{data.frame} containing all data.}
}
\description{
Generate boxplots showing the data's distance to (0/0).
}
