% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{donut}
\alias{donut}
\title{Donut perturbation mask}
\usage{
donut(p, r1, r2)
}
\arguments{
\item{p}{Original point.}

\item{r1}{Inner ring radius.}

\item{r2}{Outer ring radius.}
}
\value{
Perturbated point.
}
\description{
Adds perturbation constrained on a ring (donut) around a point.
}
