% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(n = 100, x = 0.5, y = 0.5, r = 0.1)
}
\arguments{
\item{n}{Number of points.}

\item{x}{X position of center.}

\item{y}{Y position of center}

\item{r}{Radius of the circle in which the data lies.}
}
\value{
A \code{data.frame} of x/y coordinates.
}
\description{
Generate randomised test data. The data is uniformly distributed within a
given radius.
}
