% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_descriptive_stats}
\alias{plot_descriptive_stats}
\title{Plot descriptive statistics}
\usage{
plot_descriptive_stats(stats)
}
\arguments{
\item{stats}{input data.}
}
\value{
A \code{ggplot2} object of the plot.
}
\description{
Plot the statistics calculated by descriptive_stats.
}
