% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{distance_to_zero}
\alias{distance_to_zero}
\title{Distance to zero}
\usage{
distance_to_zero(dat, xlabels)
}
\arguments{
\item{dat}{\code{list} containing all data.}

\item{xlabels}{\code{vector} of perturbation values as labels for the plot.}
}
\description{
Generate boxplots showing the data's distance to (0/0).
}
