% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_animation.R
\name{save_animated_plot}
\alias{save_animated_plot}
\title{Plot animation wrapper}
\usage{
save_animated_plot(plots, fps, file_out)
}
\arguments{
\item{plots}{\code{list} of ggplot objects.}

\item{fps}{frames per second.}

\item{file_out}{output file location.}
}
\description{
Generate an animated gif from a list of ggplot objects.
}
